bindgen: 0.53.0

+++ src/wrapper.rs
@@ -7,8 +7,9 @@ pub fn sqli(input: &str) -> Option<(bool, String)> {
     let fingerprint_raw_ptr = fingerprint_cstring.into_raw();
     let input_cstring = CString::new(input).ok()?;
     let input_ptr = input_cstring.as_ptr();
-    let is_sqli =
-        unsafe { bindings::libinjection_sqli(input_ptr, input.len(), fingerprint_raw_ptr) };
+    let is_sqli = unsafe {
+        bindings::libinjection_sqli(input_ptr, input.len() as u64, fingerprint_raw_ptr)
+    };
     Some((
         is_sqli == 1,
         unsafe { CString::from_raw(fingerprint_raw_ptr) }
@@ -21,6 +22,6 @@ pub fn sqli(input: &str) -> Option<(bool, String)> {
 pub fn xss(input: &str) -> Option<bool> {
     let input_cstring = CString::new(input).ok()?;
     let input_ptr = input_cstring.as_ptr();
-    let is_xss = unsafe { bindings::libinjection_xss(input_ptr, input.len()) };
+    let is_xss = unsafe { bindings::libinjection_xss(input_ptr, input.len() as u64) };
     Some(is_xss == 1)
 }
